requirements.txt
streamlit
PyMuPDF
langchain-anthropic
python-dotenv
edge-tts
moviepy
Pillow
numpy
pydub
SpeechRecognition